Refactor: Move ObservableArray.Observer one level up
